claim-backend/requirements.txt
blinker==1.8.2
boto3==1.35.55
botocore==1.35.55
certifi==2024.8.30
charset-normalizer==3.4.0
claim-function-library==0.1.0
claims-lib==0.1.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
docutils==0.21.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==25.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
pkginfo==1.10.0
Pygments==2.18.0
PyJWT==2.9.0
pymon==1.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
s3transfer==0.10.3
setuptools==75.5.0
six==1.16.0
twine==5.1.1
urllib3==2.2.3
Werkzeug==3.0.6
wheel==0.45.0
WTForms==3.2.1
zipp==3.21.0
